requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.7
idna==2.8
mysqlclient==1.4.2.post1
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pytz==2018.9
pyvmomi==6.7.1.2018.12
requests==2.21.0
six==1.12.0
urllib3==1.24.1
